fix: refactor test to follow an ordered execution
This helps avoid missing local registry errors
