+++ shellx-carvel/README.md
@@ -1,15 +1,19 @@
 ## Golang scripting to manage releases
 - This codebase showcases use of Golang to script a project's release
 - Note: _Release management is done by Carvel_
-- Motivation:
+
+- ## Motivation:
   - Goodness of bash coupled with correctness of Golang
   - Granular error handling
   - 100% unit tested
   - Modular code base
   - Reduced bash & make spaghetti
 
-## Usage
+## Test
 - `make`
+
+## Usage
+TODO
 
 ### References
 - https://github.com/gohugoio/hugo/blob/master/magefile.go
