+++ shellx-carvel/carvel_workflow_test.go
@@ -5,6 +5,7 @@ import "testing"
 func TestReleaseCarvelPackage(t *testing.T) {
 	tryInstallCarvelCLIs(t)
 	tryInstallKindCLI(t)
+	trySetupRegistryAsLocalDockerContainer(t)
 	tryAppBundleCreateAndPublish(t)
 	tryPackageRelease(t)
 }
