+++ shellx-carvel/docker_test.go
@@ -2,6 +2,6 @@ package shellx_carvel
 
 import "testing"
 
-func TestSetupRegistryAsLocalDockerContainer(t *testing.T) {
+func trySetupRegistryAsLocalDockerContainer(t *testing.T) {
 	requireNoErr(t, setupRegistryAsLocalDockerContainer())
 }
