+++ shellx-carvel/kind_test.go
@@ -1,0 +1,9 @@
+package shellx_carvel
+
+import "testing"
+
+func tryInstallKindCLI(t *testing.T) {
+	requireNoErr(t, installKindCLI())
+	requireNoErr(t, ls(EnvBinPathKind))
+	requireTrue(t, isNoErr(whichKind()))
+}
